fix: Support icon display in development environment
This commit adjusts the CSS for icon backgrounds to use relative paths.
In the dev environment this allows the icon selection form to display
newly added icons in the source directory, and should also work for a
production environment in the same way.

It is important that the icon paths have a hosted URI so that the final
Markdown output is able to render the icons. During development this can
be locally hosted but the Markdown output will not be usable without
changes.

This commit also limits the displayed size of icons to fit within the
container allowing larger icons to be used.